Back/requirements.txt
flask==2.3.3
pillow==10.0.0
tensorflow==2.20.0
numpy==1.26.0
scipy==1.11.3